There used to be logic to reject any input that has at least one lower case character. However, data shows that this happens way too often for valid inputs. Moreover, data also shows that user's sometimes try to pronounce words instead of providing the spelling of the word they want to pronounce.
In this commit, we remove the restriction that input cannot contain lower case
characters.
We also treat all lower case inputs (which happens typically when users try to
pronounce entire words or phrases instead of providing spelling) as a special case
and render an educational pronpt to the user.

Unit tests are adjusted accordingly.