Remove extraneous phrases from slot values.
When users say utterances like "how to pronounce the word D. O. G.",
ASK is returning 'the word dog' as slot value instead of just 'dog'.

Changing the sample utterances in ASK portal didn't help. So, we need
a solution that looks for extraneous phrases in the input and strip
them out.