Support to suggest spellings.
When a misspelling is recognized (either user fault or misrecognition
by Alexa), we create a set of suggested spellings (words with similar
spellings) and offer to pronounce them.

The user can then say 'Yes' to iterate through all the suggestions
one at a time or choose to say 'No' at any point.

This commit also includes some minor prompt changes and a couple
of typo fixes.