Fix broken tests after introducing Alexa Presentation Language.
With the introduction of APL, there is now a fork across the code based
on whether or not the Alexa endpoint supports APL. This resuled in a
bunch of broken tests (submitted the broken tests regardless in an attempt
to get the skill through Alexa certification).

In this commit, the tests are also forked based on APL support and now
pass.