APL support for error messages.
Introduced APL cards for the error message cases as well so there is a
uniform experience across the skill. Reused the APL documents used by
the HelpIntent for the error message. It isn't the best fit but the error
scenarios happen very rarely and so I figured it is not worth creating
new APL documents.